requirements.txt
atomicwrites==1.4.0
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.4
Django==2.2.6
django-debug-toolbar==3.2.2
djangorestframework==3.12.4
Faker==8.11.0
idna==2.8
importlib-metadata==1.5.0
mixer==7.1.2
more-itertools==8.2.0
packaging==20.1
Pillow==8.3.1
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pytest==5.3.5
pytest-django==3.8.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
pytz==2019.3
requests==2.22.0
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.25.6
wcwidth==0.1.8
zipp==2.2.0
